backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
Pillow
psycopg2-binary
python-dotenv
django-redis
dj-database-url
gunicorn
